PDF: preserve multi-codepoint ToUnicode clusters for text extraction
Build full CID->codepoint sequences in the PDF reader, emit logical units for clusters, and handle them in the text/Docx/txt renderers and the wasm text stream.

+++ PdfFile/SrcWriter/LogicalTrueTypeSubset.cpp
@@ -1164,7 +1164,7 @@ namespace PdfWriter
 			AppendU16(hmtx, advance);
 			AppendS16(hmtx, stats.HasContours ? stats.Bounds.XMin : 0);
 			outputStats.push_back(stats);
-			outputMetrics.push_back({advance, stats.HasContours ? stats.Bounds.XMin : 0,
+			outputMetrics.push_back({advance, static_cast<std::int16_t>(stats.HasContours ? stats.Bounds.XMin : 0),
 			                         stats.HasContours, stats.Bounds});
 		}
 		if (glyf.size() > std::numeric_limits<std::uint32_t>::max())
